Image Hoster Bug Fixes ======================
Issue #3- Non-owner of the image can edit/delete the image
Fixed this issue with proper modifications on controller and view file.

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -94,13 +94,22 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
+        String error = "Only the owner of the image can edit the image";
         Image image = imageService.getImage(imageId);
 
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        //To show error message if a non-owner try to edit an image
+        if(image.getUser().getId() == ((User)session.getAttribute("loggeduser")).getId()){
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }else{
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("editError", error);
+            return "images/image";
+        }
+
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -142,9 +151,19 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,Model model,HttpSession session) {
+        String error = "Only the owner of the image can delete the image";
+        Image image = imageService.getImageById(imageId);
+        //To show error message if a non-owner try to delete an image
+        if(image.getUser().getId() == ((User)session.getAttribute("loggeduser")).getId()){
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }else{
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        }
     }
